+++ packages/mistralai-azure/src/lib/config.ts
@@ -55,8 +55,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.2",
-  sdkVersion: "1.3.0",
-  genVersion: "2.452.0",
+  sdkVersion: "1.3.1",
+  genVersion: "2.455.2",
   userAgent:
-    "speakeasy-sdk/typescript 1.3.0 2.452.0 0.0.2 @mistralai/mistralai-azure",
+    "speakeasy-sdk/typescript 1.3.1 2.455.2 0.0.2 @mistralai/mistralai-azure",
 } as const;