+++ packages/mistralai-azure/src/models/components/chatcompletionrequest.ts
@@ -295,8 +295,8 @@ export const ChatCompletionRequest$inboundSchema: z.ZodType<
   tools: z.nullable(z.array(Tool$inboundSchema)).optional(),
   tool_choice: z.union([ToolChoice$inboundSchema, ToolChoiceEnum$inboundSchema])
     .optional(),
-  presence_penalty: z.number().default(0),
-  frequency_penalty: z.number().default(0),
+  presence_penalty: z.number().optional(),
+  frequency_penalty: z.number().optional(),
   n: z.nullable(z.number().int()).optional(),
   safe_prompt: z.boolean().default(false),
 }).transform((v) => {
@@ -330,8 +330,8 @@ export type ChatCompletionRequest$Outbound = {
   response_format?: ResponseFormat$Outbound | undefined;
   tools?: Array<Tool$Outbound> | null | undefined;
   tool_choice?: ToolChoice$Outbound | string | undefined;
-  presence_penalty: number;
-  frequency_penalty: number;
+  presence_penalty?: number | undefined;
+  frequency_penalty?: number | undefined;
   n?: number | null | undefined;
   safe_prompt: boolean;
 };
@@ -379,8 +379,8 @@ export const ChatCompletionRequest$outboundSchema: z.ZodType<
     ToolChoice$outboundSchema,
     ToolChoiceEnum$outboundSchema,
   ]).optional(),
-  presencePenalty: z.number().default(0),
-  frequencyPenalty: z.number().default(0),
+  presencePenalty: z.number().optional(),
+  frequencyPenalty: z.number().optional(),
   n: z.nullable(z.number().int()).optional(),
   safePrompt: z.boolean().default(false),
 }).transform((v) => {