+++ packages/mistralai-azure/src/models/components/index.ts
@@ -15,6 +15,7 @@ export * from "./deltamessage.js";
 export * from "./function.js";
 export * from "./functioncall.js";
 export * from "./functionname.js";
+export * from "./httpvalidationerror.js";
 export * from "./responseformat.js";
 export * from "./responseformats.js";
 export * from "./security.js";